Migrate outPt logic to out rec
