+++ packages/geometry-utils/src/clipper/out-pt.ts
@@ -30,211 +30,8 @@ export default class OutPt {
         OutPt.points.length = 0;
     }
 
-    public remove(): OutPt {
-        const result = OutPt.at(this.prev);
-        OutPt.push(this.prev, this.next, true);
-        this.prev = UNASSIGNED;
-        this.next = UNASSIGNED;
-        
-        return result;
-    }
-
-    public static duplicate(index: number,  isInsertAfter: boolean): number {
-        const outPt = OutPt.at(index);
-        const result: OutPt = new OutPt(outPt.point);
-
-        if (isInsertAfter) {
-            OutPt.push(result.current, outPt.next, true);
-            OutPt.push(index, result.current, true);
-        } else {
-            OutPt.push(outPt.prev, result.current, true);
-            OutPt.push(result.current, index, true);
-        }
-
-        return result.current;
-    }
-
     public get sameAsNext(): boolean {
         return this.current === this.next;
-    }
-
-    public static reverse(index: number): void {
-        let pp1: OutPt = OutPt.at(index);
-
-        do {
-            const pp2 = OutPt.at(pp1.next);
-            pp1.next = pp1.prev;
-            pp1.prev = pp2.current;
-            pp1 = pp2;
-        } while (pp1.current !== index);
-    }
-
-    public getUniquePt(isNext: boolean): number {
-        let result = OutPt.getNeighboarIndex(this.current, isNext);
-
-        while (OutPt.almostEqual(result, this.current) && result !== this.current) {
-            result = OutPt.getNeighboarIndex(result, isNext);
-        }
-
-        return result;
-    }
-
-    private searchBottom(outIndex1: number, outIndex2: number, isNext: boolean): number {
-        let currIndex  = this.current;
-        let nghbIndex = OutPt.getNeighboarIndex(currIndex, isNext);
-        let currPt: OutPt = OutPt.at(currIndex);
-        let nghbPt: OutPt = OutPt.at(nghbIndex);
-
-        while (nghbPt.point.y === currPt.point.y && nghbIndex !== outIndex1 && nghbIndex !== outIndex2) {
-            currIndex = nghbIndex;
-            nghbIndex = OutPt.getNeighboarIndex(currIndex, isNext);
-            currPt = OutPt.at(currIndex);
-            nghbPt = OutPt.at(nghbIndex);
-        }
-
-        return currIndex;
-    }
-
-    public flatHorizontal(outIndex1: number, outIndex2: number): number[] {
-        const outIndex = this.searchBottom(this.current, outIndex2, false);
-        const outBIndex = this.searchBottom(outIndex, outIndex1, true);
-        const outBIndexNext = OutPt.getNeighboarIndex(outBIndex, true);
-
-        return outBIndexNext === outIndex || outBIndexNext === outIndex1 ? [] : [outIndex, outBIndex];
-    }
-
-    public strictlySimpleJoin(point: Point<Int32Array>): boolean {
-        let result = OutPt.at(this.next);
-
-        while (result.current !== this.current && result.point.almostEqual(point)) {
-            result = OutPt.at(result.next);
-        }
-
-        return result.point.y > point.y;
-    }
-
-    public static applyJoin(index1: number, index2: number, reverse: boolean): number {
-        const op1b = OutPt.duplicate(index1, !reverse);
-        const op2b = OutPt.duplicate(index2, reverse);
-
-        if (reverse) {
-            OutPt.push(index2, index1, true);
-            OutPt.push(op1b, op2b, true);
-        } else {
-            OutPt.push(index1, index2, true);
-            OutPt.push(op2b, op1b, true);
-        }
-
-        return op1b;
-    }
-
-    public static pointIn(inputIndex: number, outIndex: number): number {
-        let inputPt = OutPt.at(outIndex);
-        //returns 0 if false, +1 if true, -1 if pt ON polygon boundary
-        //http://citeseerx.ist.psu.edu/viewdoc/download?doi=10.1.1.88.5498&rep=rep1&type=pdf
-        let outPt: OutPt = OutPt.at(inputIndex);
-        let startOutPt: OutPt = outPt;
-        let result: number = 0;
-        let poly0x: number = 0;
-        let poly0y: number = 0;
-        let poly1x: number = 0;
-        let poly1y: number = 0;
-        let d: number = 0;
-
-        while (true) {
-            const nextPt = OutPt.at(outPt.next);
-            poly0x = outPt.point.x;
-            poly0y = outPt.point.y;
-            poly1x = nextPt.point.x;
-            poly1y = nextPt.point.y;
-
-            if (poly1y === inputPt.point.y) {
-                if (poly1x === inputPt.point.x || (poly0y === inputPt.point.y && poly1x > inputPt.point.x === poly0x < inputPt.point.x)) {
-                    return UNASSIGNED;
-                }
-            }
-
-            if (poly0y < inputPt.point.y !== poly1y < inputPt.point.y) {
-                if (poly0x >= inputPt.point.x) {
-                    if (poly1x > inputPt.point.x) {
-                        result = 1 - result;
-                    } else {
-                        d = (poly0x - inputPt.point.x) * (poly1y - inputPt.point.y) - (poly1x - inputPt.point.x) * (poly0y - inputPt.point.y);
-
-                        if (d == 0) {
-                            return UNASSIGNED;
-                        }
-
-                        if (d > 0 === poly1y > poly0y) {
-                            result = 1 - result;
-                        }
-                    }
-                } else {
-                    if (poly1x > inputPt.point.x) {
-                        d = (poly0x - inputPt.point.x) * (poly1y - inputPt.point.y) - (poly1x - inputPt.point.x) * (poly0y - inputPt.point.y);
-
-                        if (d === 0) {
-                            return UNASSIGNED;
-                        }
-
-                        if (d > 0 === poly1y > poly0y) {
-                            result = 1 - result;
-                        }
-                    }
-                }
-            }
-
-            outPt = OutPt.at(outPt.next);
-
-            if (startOutPt.current === outPt.current) {
-                break;
-            }
-        }
-
-        return result;
-    }
-
-    public static join(index1: number, index2: number, side1: DIRECTION, side2: DIRECTION): number {
-        const prevIndex1: number = OutPt.getNeighboarIndex(index1, false);
-        const prevIndex2: number = OutPt.getNeighboarIndex(index2, false);
-        const isLeft = side1 === DIRECTION.LEFT;
-
-        if(side1 === side2) {
-            OutPt.reverse(index2);
-            OutPt.push(index2, index1, true);
-            OutPt.push(prevIndex1, prevIndex2, true);
-
-            return isLeft ? prevIndex2 : index1;
-        }
-
-        OutPt.push(prevIndex1, index2, true);
-        OutPt.push(prevIndex2, index1, true); 
-
-        return isLeft ? index2 : index1;
-    }
-
-    private static getDistance(inputIndex: number, isNext: boolean): number {
-        let index = OutPt.getNeighboarIndex(inputIndex, isNext);
-        
-        if(index === UNASSIGNED) {
-            return Number.NaN;
-        }
-
-        while (OutPt.almostEqual(inputIndex, index) && index !== inputIndex) {
-            index = OutPt.getNeighboarIndex(index, isNext);
-
-            if(index === UNASSIGNED) {
-                return Number.NaN;
-            }
-        }
-
-        const point: OutPt = OutPt.at(index);
-        const outPt = OutPt.at(inputIndex);
-        const offsetY: number = point.point.y - outPt.point.y;
-        const offsetX: number = point.point.x - outPt.point.x;
-        const result = offsetY === 0 ? HORIZONTAL : offsetX / offsetY;
-
-        return Math.abs(result);
     }
 
     public static almostEqual(index1: number, index2: number): boolean {
@@ -256,131 +53,6 @@ export default class OutPt {
         }
 
         return isNext ? outPt.next : outPt.prev;
-    }
-
-    public static getBottomPt(inputIndex: number): number {
-        let outIndex1 = inputIndex;
-        let outIndex2 = OutPt.getNeighboarIndex(inputIndex, true);
-        let dupsIndex: number = UNASSIGNED;
-
-        while (outIndex2 !== outIndex1) {
-            let outPt1: OutPt = OutPt.at(outIndex1);
-            let outPt2: OutPt = OutPt.at(outIndex2);
-
-            if (outPt2.point.y > outPt1.point.y) {
-                outIndex1 = outIndex2;
-                dupsIndex = UNASSIGNED;
-            } else if (outPt2.point.y == outPt1.point.y && outPt2.point.x <= outPt1.point.x) {
-                if (outPt2.point.x < outPt1.point.x) {
-                    dupsIndex = UNASSIGNED;
-                    outIndex1 = outIndex2;
-                } else if (outPt2.next !== outIndex1 && outPt2.prev !== outIndex1) {
-                    dupsIndex = outIndex2;
-                }
-            }
-
-            outIndex2 = outPt2.next;
-        }
-
-        if (dupsIndex !== UNASSIGNED) {
-            //there appears to be at least 2 vertices at bottomPt so ...
-            while (dupsIndex !== outIndex2) {
-                if (!OutPt.firstIsBottomPt(outIndex2, dupsIndex)) {
-                    outIndex1 = dupsIndex;
-                }
-
-                dupsIndex = OutPt.getNeighboarIndex(dupsIndex, true);
-                
-                while (!OutPt.almostEqual(dupsIndex, outIndex1)) {
-                    dupsIndex = OutPt.getNeighboarIndex(dupsIndex, true);
-                }
-            }
-        }
-
-        return outIndex1;
-    }
-
-    private static getDirection(index1: number, index2: number): DIRECTION {
-        const outPt1 = OutPt.at(index1);
-        const outPt2 = OutPt.at(index2);
-
-        return outPt1.point.x > outPt2.point.x ? DIRECTION.LEFT : DIRECTION.RIGHT
-    }
-
-
-    public static firstIsBottomPt(btmIndex1: number, btmIndex2: number): boolean {
-        const dx1p: number = OutPt.getDistance(btmIndex1, false);
-        const dx1n: number = OutPt.getDistance(btmIndex1, true);
-        const dx2p: number = OutPt.getDistance(btmIndex2, false);
-        const dx2n: number = OutPt.getDistance(btmIndex2, true);
-
-        const maxDx: number = Math.max(dx2p, dx2n);
-
-        return dx1p >= maxDx || dx1n >= maxDx;
-    }
-
-    private getDiscarded(isRight: boolean, pt: Point<Int32Array>): boolean {
-        if (this.next === UNASSIGNED) {
-            return false;
-        }
-
-        const next = OutPt.at(this.next);
-        const nextX = next.point.x;
-        const currX = this.point.x;
-        const nextY = next.point.y;
-
-        return isRight 
-            ? nextX <= pt.x && nextX >= currX && nextY === pt.y 
-            : nextX >= pt.x && nextX <= currX && nextY === pt.y;
-    } 
-
-    public static joinHorzInt(index1: number, index2: number, point: Point<Int32Array>, isDiscardLeft: boolean): { op: number, opB: number, isRightOrder: boolean } {
-        let op: OutPt = OutPt.at(index1);
-        let opB: OutPt = OutPt.at(index2);
-
-        const direction: DIRECTION = OutPt.getDirection(index1, index2);
-        const isRight = direction === DIRECTION.RIGHT;
-        const isRightOrder = isDiscardLeft !== isRight;
-
-        while (op.getDiscarded(isRight, point)) {
-            op = OutPt.at(op.next);
-        }
-
-        if (!isRightOrder && op.point.x !== point.x) {
-            op = OutPt.at(op.next);
-        }
-
-        opB = OutPt.at(OutPt.duplicate(op.current, isRightOrder));
-
-        if (!opB.point.almostEqual(point)) {
-            op = opB;
-            //op1.Pt = Pt;
-            op.point.update(point);
-            opB = OutPt.at(OutPt.duplicate(op.current, isRightOrder));
-        }
-
-        return { op: op.current, opB: opB.current, isRightOrder };
-    }
-
-    public static joinHorz(op1Index: number, op1bIndex: number, op2Index: number, op2bIndex: number, Pt: Point<Int32Array>, isDiscardLeft: boolean) {
-        const direction1: DIRECTION = OutPt.getDirection(op1Index, op1bIndex);
-        const direction2: DIRECTION = OutPt.getDirection(op2Index, op2bIndex);
-
-        if (direction1 === direction2) {
-            return false;
-        }
-        //When DiscardLeft, we want Op1b to be on the Left of Op1, otherwise we
-        //want Op1b to be on the Right. (And likewise with Op2 and Op2b.)
-        //So, to facilitate this while inserting Op1b and Op2b ...
-        //when DiscardLeft, make sure we're AT or RIGHT of Pt before adding Op1b,
-        //otherwise make sure we're AT or LEFT of Pt. (Likewise with Op2b.)
-        const join1 = OutPt.joinHorzInt(op1Index, op1bIndex, Pt, isDiscardLeft);
-        const join2 = OutPt.joinHorzInt(op2Index, op2bIndex, Pt, isDiscardLeft);
-
-        OutPt.push(join1.op, join2.op, join1.isRightOrder);
-        OutPt.push(join1.opB, join2.opB, !join1.isRightOrder);
-
-        return true;
     }
 
     public static push(outPt1Index: number, outPt2Index: number, isReverse: boolean): void {
